refactorig add notice gallary
