+++ src/pages/NoticesPage.js
@@ -1,24 +1,25 @@
 import { Helmet } from 'react-helmet';
 
-import { useEffect } from 'react';
-import { useLocation } from 'react-router-dom';
-import useAuth from '../hooks/useAuth';
+// import { useEffect } from 'react';
+// import { useLocation } from 'react-router-dom';
+// import useAuth from '../hooks/useAuth';
 // import getNoticesByCategory from '../servises/fetchNotices';
 
 import SectionContainer from '../components/SectionContainer';
 import NoticesCategoriesNav from '../components/NoticesElements/NoticesCategoriesNav/NoticesCategoriesNav';
 import NoticesSearch from '../components/NoticesElements/NoticesSearch/NoticesSearch';
-// import AddNoticeButton from '../components/NoticesElements/AddNoticeButton/AddNoticeButton';
-import NoticesCategoriesListSecond from '../components/NoticesElements/NoticesCategoriesListSecond';
-import {
-  fetchNotices,
-  fetchFavoriteNotices,
-} from '../redux/notices/operations ';
-import { selectNoticesObj } from '../redux/notices/selectors';
-import { useDispatch, useSelector } from 'react-redux';
+import NoticesGallary from '../components/NoticesElements/NoticesGallary';
+// // import AddNoticeButton from '../components/NoticesElements/AddNoticeButton/AddNoticeButton';
+// import NoticesCategoriesListSecond from '../components/NoticesElements/NoticesCategoriesListSecond';
+// import {
+//   fetchNotices,
+//   fetchFavoriteNotices,
+// } from '../redux/notices/operations ';
+// import { selectNoticesObj } from '../redux/notices/selectors';
+// import { useDispatch, useSelector } from 'react-redux';
 
-// import { setCategory } from '../redux/notices/noticesSlice';
-import Spinner from '../components/Spinner';
+// // import { setCategory } from '../redux/notices/noticesSlice';
+// import Spinner from '../components/Spinner';
 const NoticesPage = () => {
   // const [notices, setNotices] = useState([]);
   // const { pathname: category } = useLocation();
@@ -36,26 +37,26 @@ const NoticesPage = () => {
   //   fetchNotices();
   // }, [category]);
 
-  const dispatch = useDispatch();
+  // const dispatch = useDispatch();
 
   // Получаем части состояния
-  const { isLoading, error } = useSelector(selectNoticesObj);
-  // Вызываем операцию
-  const { pathname: category } = useLocation();
-  // const [state, setstate] = useState(initialState);
+  // const { isLoading, error } = useSelector(selectNoticesObj);
+  // // Вызываем операцию
+  // const { pathname: category } = useLocation();
+  // // const [state, setstate] = useState(initialState);
+
+  // // useEffect(() => {
+  // //   dispatch(fetchFavoriteNotices());
+  // // }, [dispatch]);
+  // const { isLoggedIn } = useAuth();
 
   // useEffect(() => {
-  //   dispatch(fetchFavoriteNotices());
-  // }, [dispatch]);
-  const { isLoggedIn } = useAuth();
+  //   if (isLoggedIn) {
+  //     dispatch(fetchFavoriteNotices({ category: '/notices/favorite' }));
+  //   }
 
-  useEffect(() => {
-    if (isLoggedIn) {
-      dispatch(fetchFavoriteNotices({ category: '/notices/favorite' }));
-    }
-
-    dispatch(fetchNotices({ category }));
-  }, [dispatch, category, isLoggedIn]);
+  //   dispatch(fetchNotices({ category }));
+  // }, [dispatch, category, isLoggedIn]);
 
   return (
     <>
@@ -67,12 +68,13 @@ const NoticesPage = () => {
         <NoticesSearch />
 
         <NoticesCategoriesNav />
+        <NoticesGallary />
         {/* {isLoading && <Loader />} */}
-        {error && <p>{error}</p>}
-        <div>{isLoading && 'Request in progress...'}</div>
-        {isLoading && <Spinner />}
+        {/* {error && <p>{error}</p>}
+        <div>{isLoading && 'Request in progress...'}</div> */}
+        {/* {isLoading && <Spinner />} */}
         {/* <div>{!error && !isLoading && 'Data recived...'}</div> */}
-        {!error && !isLoading && <NoticesCategoriesListSecond />}
+        {/* {!error && !isLoading && <NoticesCategoriesListSecond />} */}
         {/* <NoticesCategoriesListSecond /> */}
       </SectionContainer>
     </>
